refactor(repo): ApplicationRepository 조회/구독을 ApplicationRepositoryQueries로 분할
840줄 → 541줄 (코딩스타일 800줄 상한 준수). 읽기 전용 쿼리 8개와 Realtime
구독 2개 본문을 executeXxx/subscribeXxx 자유함수로 추출하고 클래스는 delegate만 유지.

- 추출: getById/getByApplicantId/getByApplicantIdWithStatuses/getByJobPostingId/
  hasApplied/getStatsByApplicantId/getCancellationRequests/findByJobPostingWithStats
  + subscribeByApplicantIdWithStatuses/subscribeByJobPosting
- 구독 함수 내부 this.getByApplicantIdWithStatuses → 추출된 자유함수 직접 호출
- 공유 헬퍼 rethrowOrHandle은 기존 Helpers 그대로 사용
- export 경로·메서드 시그니처 불변 → 호출부 변경 0
- 순수 이동, 기능 변경 없음
